requirements.txt
docker==3.1.1
urllib3==1.22
requests>=2.18.4
pytz>=2015.7
Django==1.11.13
djangorestframework>=3.8.2
django-pure-pagination==0.3.0
django-split-settings==0.3.0
libvirt-python==4.0
ipaddr>=2.1.11
structlog>=16.1.0
pyyaml
easydict
